+++ dev/sparktestsupport/modules.py
@@ -798,7 +798,11 @@ pyspark_pandas_slow = Module(
         "pyspark.pandas.tests.indexes.test_base",
         "pyspark.pandas.tests.indexes.test_conversion",
         "pyspark.pandas.tests.indexes.test_drop",
+        "pyspark.pandas.tests.indexes.test_level",
+        "pyspark.pandas.tests.indexes.test_missing",
+        "pyspark.pandas.tests.indexes.test_repeat",
         "pyspark.pandas.tests.indexes.test_sort",
+        "pyspark.pandas.tests.indexes.test_stat",
         "pyspark.pandas.tests.indexes.test_symmetric_diff",
         "pyspark.pandas.tests.indexes.test_take",
         "pyspark.pandas.tests.indexes.test_unique",
@@ -1094,7 +1098,11 @@ pyspark_pandas_connect_part0 = Module(
         "pyspark.pandas.tests.connect.indexes.test_parity_base",
         "pyspark.pandas.tests.connect.indexes.test_parity_conversion",
         "pyspark.pandas.tests.connect.indexes.test_parity_drop",
+        "pyspark.pandas.tests.connect.indexes.test_parity_level",
+        "pyspark.pandas.tests.connect.indexes.test_parity_missing",
+        "pyspark.pandas.tests.connect.indexes.test_parity_repeat",
         "pyspark.pandas.tests.connect.indexes.test_parity_sort",
+        "pyspark.pandas.tests.connect.indexes.test_parity_stat",
         "pyspark.pandas.tests.connect.indexes.test_parity_symmetric_diff",
         "pyspark.pandas.tests.connect.indexes.test_parity_take",
         "pyspark.pandas.tests.connect.indexes.test_parity_unique",
