+++ python/pyspark/pandas/tests/indexes/test_repeat.py
@@ -1,0 +1,65 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+import unittest
+
+import pandas as pd
+
+from pyspark import pandas as ps
+from pyspark.testing.pandasutils import PandasOnSparkTestCase
+from pyspark.testing.sqlutils import SQLTestUtils
+
+
+class RepeatMixin:
+    def test_repeat(self):
+        pidx = pd.Index(["a", "b", "c"])
+        psidx = ps.from_pandas(pidx)
+
+        self.assert_eq(psidx.repeat(3).sort_values(), pidx.repeat(3).sort_values())
+        self.assert_eq(psidx.repeat(0).sort_values(), pidx.repeat(0).sort_values())
+        self.assert_eq((psidx + "x").repeat(3).sort_values(), (pidx + "x").repeat(3).sort_values())
+
+        self.assertRaises(ValueError, lambda: psidx.repeat(-1))
+        self.assertRaises(TypeError, lambda: psidx.repeat("abc"))
+
+        pmidx = pd.MultiIndex.from_tuples([("x", "a"), ("x", "b"), ("y", "c")])
+        psmidx = ps.from_pandas(pmidx)
+
+        self.assert_eq(psmidx.repeat(3).sort_values(), pmidx.repeat(3).sort_values())
+        self.assert_eq(psmidx.repeat(0).sort_values(), pmidx.repeat(0).sort_values(), almost=True)
+
+        self.assertRaises(ValueError, lambda: psmidx.repeat(-1))
+        self.assertRaises(TypeError, lambda: psmidx.repeat("abc"))
+
+
+class RepeatTests(
+    RepeatMixin,
+    PandasOnSparkTestCase,
+    SQLTestUtils,
+):
+    pass
+
+
+if __name__ == "__main__":
+    from pyspark.pandas.tests.indexes.test_repeat import *  # noqa: F401
+
+    try:
+        import xmlrunner
+
+        testRunner = xmlrunner.XMLTestRunner(output="target/test-reports", verbosity=2)
+    except ImportError:
+        testRunner = None
+    unittest.main(testRunner=testRunner, verbosity=2)
